feat: add `reloadConfigOnSave` as config value
