+++ server/src/configs/default.json
@@ -1,6 +1,7 @@
 {
   "interface": "0.0.0.0",
   "port": 8080,
+  "reloadConfigOnSave": false,
   "packageSource": "git",
   "googleAnalyticsId": "",
   "devOptions": {
